refactor: Phase 1 — split xstateUtilsV5.ts into focused modules
Split the 2,001-line God Object into 4 focused modules:

- xstateTypes.ts (14 lines) — PersistedXStateData, PreApprovalUpdateData
- xstateEffects.ts (781 lines) — handleStateTransitions, sendCanceledEmail
- xstatePersistence.ts (569 lines) — snapshot save/restore/migration,
  machine selection, status mapping
- xstateTransitions.ts (667 lines) — executeXStateTransition,
  getAvailableXStateTransitions

xstateUtilsV5.ts becomes a 1-line barrel re-export so external callers
are unaffected.

No runtime behavior changes. All 83 test files (1,318 tests) pass.
